fix: cast explicitly to long double

+++ include/raytracing_scene.hpp
@@ -120,12 +120,13 @@ class RaytracingScene
 			    = ltracer::extractBezierTriangleFromTetrahedron(tetrahedron2, side);
 
 			std::vector<BezierTriangle2> subTriangles;
-			subTriangles.reserve(static_cast<size_t>(std::powl(4, subdivisions)));
+			subTriangles.reserve(
+			    static_cast<size_t>(std::powl(4.0L, static_cast<long double>(subdivisions))));
 			subTriangles.push_back(bezierTriangle);
 
 			for (int i = 0; i < subdivisions; i++)
 			{
-				for (size_t j = 0; j < std::powl(4, i); j++)
+				for (size_t j = 0; j < std::powl(4.0L, static_cast<long double>(i)); j++)
 				{
 					BezierTriangle2& t = subTriangles.front();
 					const auto& s = ltracer::subdivideBezierTriangle2(t);
